fix linter errors
