+++ src/index.js
@@ -1,6 +1,7 @@
 import './style.css';
 import { getTodosFromStorage } from './modules/storage.js';
-import { addNewTodo, removeTodo } from './modules/todoCrud.js';
+import { removeTodo } from './modules/todoUtils.js';
+import addNewTodo from './modules/newTodo.js';
 import displayTodos from './modules/display.js';
 
 // CRUD functions for To Do list
